% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary4mice-package.R
\docType{package}
\name{gtsummary4mice-package}
\alias{gtsummary4mice}
\alias{gtsummary4mice-package}
\title{gtsummary4mice: `tbl_univsummary` method for `mice::mids` objects}
\description{
Create univariate analyses tables for `mice::mids`, imputed data objects generated using `mice::mice`. Adds S3 method to the `gtsummary` summary package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jrob95/gtsummary4mice}
  \item Report bugs at \url{https://github.com/jrob95/gtsummary4mice/issues}
}

}
\author{
\strong{Maintainer}: Jack Roberts \email{jack.roberts@connect.qut.edu.au} (\href{https://orcid.org/0000-0003-3519-4725}{ORCID})

Authors:
\itemize{
  \item Daniel D. Sjoberg (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})
}

}
\keyword{internal}
